Move `LangError` integration tests into single workspace
