+++ examples/lang_err_integration_tests/call_builder/Cargo.toml
@@ -1,24 +1,25 @@
 [package]
-name = "integration_flipper"
+name = "call_builder"
 version = "4.0.0-alpha.3"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2021"
 publish = false
 
 [dependencies]
-ink = { path = "../../crates/ink", default-features = false }
+ink = { path = "../../../crates/ink", default-features = false }
 
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
 
+[dev-dependencies]
+ink_e2e = { path = "../../../crates/e2e" }
+
 [lib]
-name = "integration_flipper"
+name = "call_builder"
 path = "lib.rs"
 crate-type = [
-    # Used for normal contract Wasm blobs.
-    "cdylib",
-    # Used for ABI generation.
-    "rlib",
+	# Used for normal contract Wasm blobs.
+	"cdylib",
 ]
 
 [features]
@@ -29,9 +30,4 @@ std = [
     "scale-info/std",
 ]
 ink-as-dependency = []
-
-[workspace]
-members = [
-    "call_builder",
-    "contract_ref",
-]
+e2e-tests = []
