+++ examples/lang_err_integration_tests/contract_ref/Cargo.toml
@@ -1,23 +1,26 @@
 [package]
-name = "cross_chain_ref"
+name = "contract_ref"
 version = "0.1.0"
 authors = ["[your_name] <[your_email]>"]
 edition = "2021"
 
 [dependencies]
-ink = { path = "../../crates/ink", default-features = false }
+ink = { path = "../../../crates/ink", default-features = false }
 
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2", default-features = false, features = ["derive"], optional = true }
 
-flipper = { path = "../flipper", default-features = false, features = ["ink-as-dependency"] }
+integration_flipper = { path = "../", default-features = false, features = ["ink-as-dependency"] }
+
+[dev-dependencies]
+ink_e2e = { path = "../../../crates/e2e" }
 
 [lib]
-name = "cross_chain_ref"
+name = "contract_ref"
 path = "lib.rs"
 crate-type = [
-	# Used for normal contract Wasm blobs.
-	"cdylib",
+    # Used for normal contract Wasm blobs.
+    "cdylib",
 ]
 
 [features]
@@ -27,6 +30,7 @@ std = [
     "scale/std",
     "scale-info/std",
 
-    "flipper/std",
+    "integration_flipper/std",
 ]
 ink-as-dependency = []
+e2e-tests = []
